+++ SuperBuild/External_Gadgetron.cmake
@@ -25,7 +25,7 @@
 set(proj Gadgetron)
 
 # Set dependency list
-set(${proj}_DEPENDENCIES "Boost;HDF5;ISMRMRD;FFTW3double;Armadillo;GTest")
+set(${proj}_DEPENDENCIES "ACE;Boost;HDF5;ISMRMRD;FFTW3double;Armadillo;GTest")
 
 # Include dependent projects if any
 ExternalProject_Include_Dependencies(${proj} DEPENDS_VAR ${proj}_DEPENDENCIES)
@@ -45,26 +45,35 @@ if(NOT ( DEFINED "USE_SYSTEM_${externalProjName}" AND "${USE_SYSTEM_${externalPr
   ### --- Project specific additions here
   set(Gadgetron_Install_Dir ${SUPERBUILD_INSTALL_DIR})
 
+  # Gadgetron only adds tests if (GTEST_FOUND AND ARMADILLO_FOUND)
+  # but that's currently always the case.
+  # Default to on, as we cannot disable it, and they're quite fast
+  option(BUILD_TESTING_${proj} "Build tests for ${proj}" ON)
+
   #message(STATUS "HDF5_ROOT in External_SIRF: " ${HDF5_ROOT})
   set(CMAKE_LIBRARY_PATH ${CMAKE_LIBRARY_PATH} ${SUPERBUILD_INSTALL_DIR})
   set(CMAKE_INCLUDE_PATH ${CMAKE_INCLUDE_PATH} ${SUPERBUILD_INSTALL_DIR})
 
-  # work-around Gadgetron problem with old MKL
-find_package(MKL)
-if (MKL_FOUND)
-  if ( MKL_VERSION_STRING VERSION_LESS 11.2.0 )
-    message(WARNING "Gadgetron requires Intel MKL version >= 11.2.0 but found ${MKL_VERSION_STRING}. Disabling usage of MKL")
-    set(MKL_FOUND false)
-    set(MKLROOT_PATH=)
-  endif ()
-endif ()
+  # BLAS
+  find_package(BLAS)
+  if (APPLE AND NOT (CBLAS_LIBRARY AND CBLAS_INCLUDE_DIR))
+    message(FATAL_ERROR "Gadgetron needs CBLAS_LIBRARY and CBLAS_INCLUDE_DIR. If
+      these variables do not exist in your CMake, create them manually. CBLAS_LIBRARY
+      and CBLAS_INCLUDE_DIR should be FILEPATH and PATH, respectively, and live in
+      /usr/local/Cellar/openblas/ if installed with \"brew install openblas\".")
+  endif()
 
-  #option(BUILD_GADGETRON_NATIVE_PYTHON_SUPPORT
+  #option(Gadgetron_BUILD_PYTHON_SUPPORT
   #  "Build Gadgetron Python gadgets (not required for SIRF)" OFF)
-  set(BUILD_GADGETRON_NATIVE_PYTHON_SUPPORT OFF) # <-Disabled for v1.0
-  option(BUILD_GADGETRON_NATIVE_MATLAB_SUPPORT
-    "Build Gadgetron MATLAB gadgets (not required for SIRF)" OFF)
+  set(Gadgetron_BUILD_PYTHON_SUPPORT OFF) # <-Disabled for v1.0
+  set(default_Gadgetron_BUILD_MATLAB_SUPPORT OFF)
+  RenameVariable(BUILD_GADGETRON_MATLAB_SUPPORT Gadgetron_BUILD_MATLAB_SUPPORT default_Gadgetron_BUILD_MATLAB_SUPPORT)
+  option(Gadgetron_BUILD_MATLAB_SUPPORT
+    "Build Gadgetron MATLAB gadgets (not required for SIRF)" ${default_Gadgetron_BUILD_MATLAB_SUPPORT})
+  option(Gadgetron_USE_MKL "Instruct Gadgetron to build linking to the MKL. The user must be able to install MKL on his own." OFF)
 
+  option(Gadgetron_USE_CUDA "Enable Gadgetron CUDA (if cuda libraries are present)" ON)
+  mark_as_advanced(Gadgetron_USE_CUDA)
   
 
   ExternalProject_Add(${proj}
@@ -78,11 +87,11 @@ endif ()
     TMP_DIR ${${proj}_TMP_DIR}
 	
     CMAKE_ARGS
-        -DBUILD_PYTHON_SUPPORT=${BUILD_GADGETRON_NATIVE_PYTHON_SUPPORT}
-        -DBUILD_MATLAB_SUPPORT=${BUILD_GADGETRON_NATIVE_MATLAB_SUPPORT}
+        -DBUILD_PYTHON_SUPPORT=${Gadgetron_BUILD_PYTHON_SUPPORT}
+        -DBUILD_MATLAB_SUPPORT=${Gadgetron_BUILD_MATLAB_SUPPORT}
         -DCMAKE_PREFIX_PATH=${SUPERBUILD_INSTALL_DIR}
         -DCMAKE_LIBRARY_PATH=${SUPERBUILD_INSTALL_DIR}/lib
-        -DCMAKE_INCLUDE_PATH=${SUPERBUILD_INSTALL_DIR}
+        -DCMAKE_INCLUDE_PATH=${SUPERBUILD_INSTALL_DIR}/include
         -DCMAKE_INSTALL_PREFIX=${Gadgetron_Install_Dir}
         -DBOOST_INCLUDEDIR=${BOOST_ROOT}/include/
         -DBOOST_LIBRARYDIR=${BOOST_LIBRARY_DIR}
@@ -90,12 +99,13 @@ endif ()
         -DPYTHON_INCLUDE_DIR=${PYTHON_INCLUDE_DIRS}
         -DPYTHON_LIBRARY=${PYTHON_LIBRARIES}
         -DGTEST_ROOT=${GTEST_ROOT}
-        -DHDF5_ROOT=${HDF5_ROOT}
-        -DHDF5_INCLUDE_DIRS=${HDF5_INCLUDE_DIRS}
-        -DHDF5_LIBRARIES=${HDF5_LIBRARIES}
-        -DFFTW_ROOT_DIR=${SUPERBUILD_INSTALL_DIR}
+        ${HDF5_CMAKE_ARGS}
         -DISMRMRD_DIR=${ISMRMRD_DIR}
-        -DMKLROOT_PATH=${MKLROOT_PATH}
+	-DUSE_MKL:BOOL=${Gadgetron_USE_MKL}
+        -DUSE_CUDA=${Gadgetron_USE_CUDA}
+        -DCBLAS_INCLUDE_DIR:PATH=${CBLAS_INCLUDE_DIR}
+        -DCBLAS_LIBRARY:FILEPATH=${CBLAS_LIBRARY}
+
 	    INSTALL_DIR ${Gadgetron_Install_Dir}
     DEPENDS
         ${${proj}_DEPENDENCIES}
@@ -104,10 +114,17 @@ endif ()
     set(Gadgetron_ROOT        ${Gadgetron_SOURCE_DIR})
     set(Gadgetron_INCLUDE_DIR ${Gadgetron_SOURCE_DIR})
 
+  # Gadgetron only adds tests if (GTEST_FOUND AND ARMADILLO_FOUND)
+  if (BUILD_TESTING_${proj})
+    add_test(NAME ${proj}_TESTS
+         COMMAND ${CMAKE_CTEST_COMMAND} -C $<CONFIGURATION> --output-on-failure
+         WORKING_DIRECTORY ${${proj}_BINARY_DIR}/test)
+  endif()
+     
    else()
       if(${USE_SYSTEM_${externalProjName}})
         find_package(${proj} ${${externalProjName}_REQUIRED_VERSION} REQUIRED)
-        message("USING the system ${externalProjName}, set ${externalProjName}_DIR=${${externalProjName}_DIR}")
+        message(STATUS "USING the system ${externalProjName}, set ${externalProjName}_DIR=${${externalProjName}_DIR}")
     endif()
   ExternalProject_Add_Empty(${proj} DEPENDS "${${proj}_DEPENDENCIES}"
     SOURCE_DIR ${${proj}_SOURCE_DIR}