+++ docker/devel/test.sh
@@ -14,15 +14,20 @@ pip install -U -r "$this"/requirements-test.txt
 
 pushd $SIRF_PATH/../..
 
-./INSTALL/bin/gadgetron >& gadgetron.log&
-# print for debugging
-[ "$DEBUG" != 0 ] && cat builds/SIRF/build/CMakeCache.txt
+# start gadgetron
+GCONFIG=./INSTALL/share/gadgetron/config/gadgetron.xml
+[ -f "$GCONFIG" ] || cp "$GCONFIG".example "$GCONFIG"
+[ -f ./INSTALL/bin/gadgetron ] && ./INSTALL/bin/gadgetron >& gadgetron.log&
+
 ctest -VV
 ret=$?
+[ -n "$(pidof gadgetron)" ] && kill -n 15 $(pidof gadgetron)
+
 # print for debugging
+[ "$DEBUG" != 0 ] && cat builds/SIRF/build/CMakeCache.txt
 [ "$DEBUG" != 0 ] && cat builds/SIRF/build/Testing/Temporary/LastTest.log
 # may exceed 4MB travis log limit
-[ "$DEBUG" != 0 ] && cat gadgetron.log
+[ "$DEBUG" != 0 ] && [ -f gadgetron.log ] && cat gadgetron.log
 
 popd