+++ version_config.cmake
@@ -22,10 +22,24 @@
 #=========================================================================
 
 ## BOOST
-set(Boost_VERSION 1.63.0)
-set(Boost_REQUIRED_VERSION 1.36.0)
-set(Boost_URL http://downloads.sourceforge.net/project/boost/boost/1.63.0/boost_1_63_0.zip)
-set(Boost_MD5 3c706b3fc749884ea5510c39474fd732)
+if (APPLE) # really should be checking for CLang
+    # Boost 1.65 contains a bug for recent Clang https://github.com/CCPPETMR/SIRF-SuperBuild/issues/170
+    set(Boost_VERSION 1.68.0)
+    set(Boost_REQUIRED_VERSION 1.66.0)
+    set(Boost_URL http://downloads.sourceforge.net/project/boost/boost/${Boost_VERSION}/boost_1_68_0.zip)
+    set(Boost_MD5 f4096c4583947b0eb103c8539f1623a3)
+else()
+     # Use version in Ubuntu 18.04
+     set(Boost_VERSION 1.65.1)
+     if (BUILD_GADGETRON)
+       set(Boost_REQUIRED_VERSION 1.65.1)
+     else()
+       # Ubutnu 16.04 version should be fine
+       set(Boost_REQUIRED_VERSION 1.58.0)
+     endif()
+     set(Boost_URL http://downloads.sourceforge.net/project/boost/boost/${Boost_VERSION}/boost_1_65_1.zip)
+     set(Boost_MD5 9824a7a3e25c9d4fdf2def07bce8651c)
+endif()
 
 ## Armadillo
 set(Armadillo_URL   https://downloads.sourceforge.net/project/arma/armadillo-7.800.2.tar.xz)
@@ -51,12 +65,34 @@ set(FFTW3double_MD5 ${FFTW3_MD5})
 if (WIN32)
   # 1.8.15 hdf5-targets.cmake refers to non-existent zlib files
   # (or at least this was the case for older Anaconda installations)
-  set(HDF5_REQUIRED_VERSION 1.8.17)
+  if (USE_SYSTEM_HDF5)
+    set(HDF5_REQUIRED_VERSION 1.8.17)
+    #set(HDF5_REQUIRED_VERSION 1.8)
+  endif()
+  if (BUILD_MATLAB)
+    set(HDF5_DOWNLOAD_MAJOR_MINOR_VERSION "1.8")
+    set(HDF5_DOWNLOAD_PATCH_VERSION "12")
+  else()
+    set(HDF5_DOWNLOAD_MAJOR_MINOR_VERSION "1.10")
+    set(HDF5_DOWNLOAD_PATCH_VERSION "1")
+  endif()
+  set(HDF5_DOWNLOAD_VERSION "${HDF5_DOWNLOAD_MAJOR_MINOR_VERSION}.${HDF5_DOWNLOAD_PATCH_VERSION}")
+  set(HDF5_URL https://support.hdfgroup.org/ftp/HDF5/releases/hdf5-${HDF5_DOWNLOAD_MAJOR_MINOR_VERSION}/hdf5-${HDF5_DOWNLOAD_VERSION}/src/hdf5-${HDF5_DOWNLOAD_VERSION}.tar.gz)
 else()
   set(HDF5_REQUIRED_VERSION 1.8)
+  set(HDF5_DOWNLOAD_MAJOR_MINOR_VERSION "1.10")
+  set(HDF5_DOWNLOAD_PATCH_VERSION "1")
+  set(HDF5_DOWNLOAD_VERSION "${HDF5_DOWNLOAD_MAJOR_MINOR_VERSION}.${HDF5_DOWNLOAD_PATCH_VERSION}")
+  set(HDF5_URL https://www.ccppetmr.ac.uk/sites/www.ccppetmr.ac.uk/files/downloads/hdf5-${HDF5_DOWNLOAD_VERSION}.tar.gz)
 endif()
-set(HDF5_URL https://www.ccppetmr.ac.uk/sites/www.ccppetmr.ac.uk/files/downloads/hdf5-1.10.1.tar.gz)
-set(HDF5_MD5 43a2f9466702fb1db31df98ae6677f15 )
+
+if (${HDF5_DOWNLOAD_VERSION} STREQUAL 1.8.12)
+  set(HDF5_MD5 4711e0d6b87d442f194c5d29ba7fbe62 )
+elseif (${HDF5_DOWNLOAD_VERSION} STREQUAL 1.10.1)
+  set(HDF5_MD5 43a2f9466702fb1db31df98ae6677f15 )
+else()
+  message(FATAL_ERROR "Developer error: need to set HDF5_MD5 for version ${HDF5_DOWNLOAD_VERSION}")
+endif()
 
 ## SWIG
 set (SWIG_REQUIRED_VERSION 2)
@@ -79,7 +115,20 @@ set(glog_TAG v035)
 
 ## ITK
 set(ITK_URL https://itk.org/ITK.git)
-set(ITK_TAG v4.13.0)
+set(ITK_TAG v4.13.1)
+
+## NIFTYREG
+set(DEFAULT_NIFTYREG_URL https://github.com/KCL-BMEIS/niftyreg.git )
+set(DEFAULT_NIFTYREG_TAG 99d584e2b8ea0bffe7e65e40c8dc818751782d92 )
+set(DEFAULT_NIFTYREG_REQUIRED_VERSION 1.5.68)
+
+## ISMRMRD
+set(DEFAULT_ISMRMRD_URL https://github.com/ismrmrd/ismrmrd )
+set(DEFAULT_ISMRMRD_TAG v1.4.0)
+
+## Gadgetron
+set(DEFAULT_Gadgetron_URL https://github.com/gadgetron/gadgetron )
+set(DEFAULT_Gadgetron_TAG b6191eaaa72ccca6c6a5fe4c0fa3319694f512ab)
 
 option (DEVEL_BUILD "Developer Build" OFF)
 mark_as_advanced(DEVEL_BUILD)
@@ -90,61 +139,48 @@ mark_as_advanced(DEVEL_BUILD)
 
 set(DEFAULT_SIRF_URL https://github.com/CCPPETMR/SIRF )
 if (DEVEL_BUILD)
+
   set (DEFAULT_SIRF_TAG origin/master)
   ## STIR
   set(DEFAULT_STIR_URL https://github.com/UCL/STIR )
   set(DEFAULT_STIR_TAG origin/master)
 
-  ## Gadgetron
-  set(DEFAULT_Gadgetron_URL https://github.com/gadgetron/gadgetron )
-  set(DEFAULT_Gadgetron_TAG origin/master)
-
   ## siemens_to_ismrmrd
   set(DEFAULT_siemens_to_ismrmrd_URL https://github.com/ismrmrd/siemens_to_ismrmrd )
   set(DEFAULT_siemens_to_ismrmrd_TAG origin/master)
 
-  ## ISMRMRD
-  set(DEFAULT_ISMRMRD_URL https://github.com/ismrmrd/ismrmrd )
-  set(DEFAULT_ISMRMRD_TAG origin/master)
-
-  ## petmr-rd-tools
-  set(DEFAULT_petmr_rd_tools_URL https://github.com/UCL/petmr-rd-tools )
-  set(DEFAULT_petmr_rd_tools_TAG origin/master)
+  ## pet-rd-tools
+  set(DEFAULT_pet_rd_tools_URL https://github.com/UCL/pet-rd-tools )
+  set(DEFAULT_pet_rd_tools_TAG origin/master)
 
   ## glog
   set(DEFAULT_glog_URL https://github.com/google/glog )
   set(DEFAULT_glog_TAG v035)
+ 
+  set(DEFAULT_ACE_URL https://github.com/paskino/libace-conda)
+  set(DEFAULT_ACE_TAG origin/master)
 
 else()
-  set(DEFAULT_SIRF_TAG v1.0.0)
+  set(DEFAULT_SIRF_TAG v2.0.0)
 
   ## STIR
   set(DEFAULT_STIR_URL https://github.com/UCL/STIR )
-  set(DEFAULT_STIR_TAG 5710d438090a972bd7b0636ff6de5315519b78e9 )
-
-  ## Gadgetron
-  set(DEFAULT_Gadgetron_URL https://github.com/gadgetron/gadgetron )
-  #https://github.com/CCPPETMR/gadgetron) 
-
-  set(DEFAULT_Gadgetron_TAG e7eb430673eb3272e8a821b51750c0a2a96dafed )
-  #set(DEFAULT_Gadgetron_TAG 00b96376568278a595e78879026bb3b0d5fbb98d )
+  set(DEFAULT_STIR_TAG 3a277f7a819f35a553a8d6097402ea25cf55a240)
 
   ## siemens_to_ismrmrd
   set(DEFAULT_siemens_to_ismrmrd_URL https://github.com/ismrmrd/siemens_to_ismrmrd)
   set(DEFAULT_siemens_to_ismrmrd_TAG ba4773f9cf4bba5f3ccd19930e3548d8273fee01)
 
-  ## ISMRMRD
-  set(DEFAULT_ISMRMRD_URL https://github.com/ismrmrd/ismrmrd )
-  set(DEFAULT_ISMRMRD_TAG 42d93137cc16c270c8ba065edd2496483161bd21)
-
-  ## petmr-rd-tools
-  set(DEFAULT_petmr_rd_tools_URL https://github.com/UCL/petmr-rd-tools )
-  set(DEFAULT_petmr_rd_tools_TAG b88281f79e8c4a3781ebda7663f1ce7f5cab6e68)
+  ## pet-rd-tools
+  set(DEFAULT_pet_rd_tools_URL https://github.com/UCL/pet-rd-tools )
+  set(DEFAULT_pet_rd_tools_TAG b88281f79e8c4a3781ebda7663f1ce7f5cab6e68)
 
   ## glog
   set(DEFAULT_glog_URL https://github.com/google/glog )
   set(DEFAULT_glog_TAG v035)
 
+  set(DEFAULT_ACE_URL https://github.com/paskino/libace-conda)
+  set(DEFAULT_ACE_TAG origin/master)
 endif()
 
 
@@ -165,15 +201,22 @@ SET(siemens_to_ismrmrd_URL ${DEFAULT_siemens_to_ismrmrd_URL} CACHE STRING ON)
 SET(ISMRMRD_TAG ${DEFAULT_ISMRMRD_TAG} CACHE STRING ON)
 SET(ISMRMRD_URL ${DEFAULT_ISMRMRD_URL} CACHE STRING ON)
 
-SET(petmr_rd_tools_TAG ${DEFAULT_petmr_rd_tools_TAG} CACHE STRING ON)
-SET(petmr_rd_tools_URL ${DEFAULT_petmr_rd_tools_URL} CACHE STRING ON)
+SET(pet_rd_tools_TAG ${DEFAULT_pet_rd_tools_TAG} CACHE STRING ON)
+SET(pet_rd_tools_URL ${DEFAULT_pet_rd_tools_URL} CACHE STRING ON)
 
 SET(glog_URL ${DEFAULT_glog_URL} CACHE STRING ON)
 SET(glog_TAG ${DEFAULT_glog_TAG} CACHE STRING ON)
+
+set(ACE_URL ${DEFAULT_ACE_URL} CACHE STRING ON)
+set(ACE_TAG ${DEFAULT_ACE_TAG} CACHE STRING ON)
+
+set(NIFTYREG_URL ${DEFAULT_NIFTYREG_URL} CACHE STRING ON)
+set(NIFTYREG_TAG ${DEFAULT_NIFTYREG_TAG} CACHE STRING ON)
 
 mark_as_advanced(SIRF_URL SIRF_TAG STIR_URL STIR_TAG
   Gadgetron_URL Gadgetron_TAG
   siemens_to_ismrmrd_URL siemens_to_ismrmrd_TAG
   ISMRMRD_URL ISMRMRD_TAG
-  petmr_rd_tools_URL petmr_rd_tools_TAG
-  glog_URL glog_TAG)
+  pet_rd_tools_URL pet_rd_tools_TAG
+  glog_URL glog_TAG
+  NIFTYREG_URL NIFTYREG_TAG)